timer: Sums the number of timer counts
The number of counts is a lean way to measure time.

To measure processing time set the TIMER_PRESCALER to a low value and use
the function timerGetCounts().

There are special definitions for low prescalers, which make the function
micros() run slower. Low prescalers also increase system load due to the
frequent interrupts (8U is not that bad though).

If necessary, the conversion from counts to microsseconds will happen with
interrupts enabled.